♻️ JE-36-refactor: Banner 폴더 생성 및 Image Button, counter 컴포넌트 분리
