+++ src/components/ActivityDetail/Banner/BlurBannerImage.tsx
