+++ src/components/ActivityDetail/Banner/MobileBannerImage.tsx
@@ -1,8 +1,9 @@
-import { ArrowLeft, ArrowRight } from '@/libs/utils/Icon';
-import React, { useEffect, useRef, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 
+import ImageCounter from '../ImageCounter';
+import ImageNavButton from '../ImageNavButton';
+import ThumbnailImage from '../ThumbnailImage';
 import BlurBannerImage from './BlurBannerImage';
-import ThumbnailImage from './ThumbnailImage';
 
 interface MobileImageProps {
   bannerImageUrl: string;
@@ -25,7 +26,6 @@ export default function MobileBannerImage(props: MobileImageProps) {
     setCurrentIndex((prevIndex) => (prevIndex - 1 + totalImages) % totalImages);
   };
 
-  // 현재 인덱스를 기준으로 썸네일 스크롤 위치 조정
   useEffect(() => {
     if (thumbnailRef.current) {
       const thumbnailWidth = thumbnailRef.current.clientWidth / totalImages;
@@ -35,7 +35,7 @@ export default function MobileBannerImage(props: MobileImageProps) {
   }, [currentIndex, totalImages]);
 
   return (
-    <div className="my-16 flex flex-col">
+    <div className="my-16 flex flex-col md:hidden">
       <div
         className={`relative flex h-310 w-full items-center justify-center overflow-hidden ${className}`}
       >
@@ -43,42 +43,21 @@ export default function MobileBannerImage(props: MobileImageProps) {
           src={images[currentIndex]}
           alt={`image-${currentIndex}`}
         />
-
-        {/* 이전 이미지 버튼 */}
-        <button
-          className="absolute left-0 top-1/2 -translate-y-1/2 transform rounded bg-custom-gray-300 px-4 py-16 opacity-50"
-          onClick={prevImage}
-        >
-          <ArrowLeft className="size-20" />
-        </button>
-
-        {/* 다음 이미지 버튼 */}
-        <button
-          className="absolute right-0 top-1/2 -translate-y-1/2 transform rounded bg-custom-gray-300 px-4 py-16 opacity-50"
-          onClick={nextImage}
-        >
-          <ArrowRight className="size-20" />
-        </button>
-
-        {/* 이미지 개수 표시 버튼 */}
-        <div className="absolute bottom-10 right-16 flex">
-          <div className="rounded bg-custom-gray-700 px-8 py-2 text-custom-gray-100">
-            {`${currentIndex + 1} / ${totalImages}`}
-          </div>
-        </div>
+        <ImageNavButton direction="left" onClick={prevImage} />
+        <ImageNavButton direction="right" onClick={nextImage} />
+        <ImageCounter currentIndex={currentIndex} totalImages={totalImages} />
       </div>
-
-      {/* 썸네일 이미지 표시 */}
       <div
         ref={thumbnailRef}
         className="mx-16 mt-16 flex cursor-pointer space-x-8 overflow-x-hidden"
       >
         {images.map((img, index) => (
-          <div className="w-60 flex-none" key={index}>
+          <div className="mt-10 h-80 w-60 flex-none" key={index}>
             <ThumbnailImage
               src={img}
               alt={`thumbnail-${index}`}
-              onClick={() => setCurrentIndex(index)} // 썸네일 클릭 시 해당 이미지로 변경
+              onClick={() => setCurrentIndex(index)}
+              className={`${currentIndex === index ? 'rounded-lg outline outline-2 outline-custom-blue-300' : ''}`}
             />
           </div>
         ))}
