+++ src/components/ActivityDetail/index.tsx
@@ -1,6 +1,6 @@
-export { default as Title } from './Title';
-export { default as BannerImage } from './BannerImage';
-export { default as MobileBannerImage } from './MobileBannerImage';
-export { default as Description } from './Description';
-export { default as Address } from './Address';
+export { default as Title } from './DetailContent/Title';
+export { default as BannerImage } from './Banner/BannerImage';
+export { default as MobileBannerImage } from './Banner/MobileBannerImage';
+export { default as Description } from './DetailContent/Description';
+export { default as Address } from './DetailContent/Address';
 export { default as Review } from './Reivew/Review';
